Use dynamic quantization

+++ src/Network.h
@@ -13,7 +13,7 @@ class Position;
 
 struct HalfAccumulator
 {
-    std::array<std::array<float, HIDDEN_NEURONS>, N_PLAYERS> side;
+    std::array<std::array<int16_t, HIDDEN_NEURONS>, N_PLAYERS> side;
 };
 
 class Network
@@ -38,8 +38,8 @@ public:
 private:
     std::vector<HalfAccumulator> AccumulatorStack;
 
-    static std::array<std::array<float, HIDDEN_NEURONS>, INPUT_NEURONS> hiddenWeights;
-    static std::array<float, HIDDEN_NEURONS> hiddenBias;
-    static std::array<float, HIDDEN_NEURONS * 2> outputWeights;
-    static float outputBias;
+    static std::array<std::array<int16_t, HIDDEN_NEURONS>, INPUT_NEURONS> hiddenWeights;
+    static std::array<int16_t, HIDDEN_NEURONS> hiddenBias;
+    static std::array<int16_t, HIDDEN_NEURONS * 2> outputWeights;
+    static int16_t outputBias;
 };
